Moved models Forecast.kt and UpdateTime.kt to another module
