+++ app/src/test/java/com/gsrg/luasdublin/ui/fragments/forecast/ForecastViewModelTest.kt
@@ -1,8 +1,8 @@
 package com.gsrg.luasdublin.ui.fragments.forecast
 
+import com.gsrg.luasdublin.core.models.Forecast
+import com.gsrg.luasdublin.core.models.UpdateTime
 import com.gsrg.luasdublin.core.utils.Result
-import com.gsrg.luasdublin.database.forecast.Forecast
-import com.gsrg.luasdublin.database.updatetime.UpdateTime
 import com.gsrg.luasdublin.domain.repository.IForecastRepository
 import com.gsrg.luasdublin.utils.ICalendar
 import kotlinx.coroutines.flow.Flow
