+++ domain/src/main/java/com/gsrg/luasdublin/domain/repository/IForecastRepository.kt
@@ -1,8 +1,8 @@
 package com.gsrg.luasdublin.domain.repository
 
+import com.gsrg.luasdublin.core.models.Forecast
+import com.gsrg.luasdublin.core.models.UpdateTime
 import com.gsrg.luasdublin.core.utils.Result
-import com.gsrg.luasdublin.database.forecast.Forecast
-import com.gsrg.luasdublin.database.updatetime.UpdateTime
 import kotlinx.coroutines.flow.Flow
 
 interface IForecastRepository {
